+++ test/test_source_pil.py
@@ -78,8 +78,8 @@ def testGetBandInformation():
     source = large_image_source_pil.open(imagePath)
     bandInfo = source.getBandInformation(False)
     assert len(bandInfo) == 4
-    assert bandInfo[0] == {'interpretation': 'red'}
+    assert bandInfo[1] == {'interpretation': 'red'}
 
     bandInfo = source.getBandInformation(True)
     assert len(bandInfo) == 4
-    assert 'mean' in bandInfo[0]
+    assert 'mean' in bandInfo[1]