+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Bug Fixes
 - getBandInformation could fail on high bands in some cases (#651)
+- Band information should always be 1-indexed (#659)
 
 ## Version 1.8.0